+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ux/environment-check
 
+## 0.19.2
+
+### Patch Changes
+
+- Updated dependencies [9752c40]
+    - @sap-ux/ui5-config@0.31.1
+    - @sap-ux/project-access@1.38.1
+    - @sap-ux/axios-extension@1.26.0
+
+## 0.19.1
+
+### Patch Changes
+
+- Updated dependencies [63e6846]
+    - @sap-ux/project-access@1.38.0
+    - @sap-ux/axios-extension@1.26.0
+
 ## 0.19.0
 
 ### Minor Changes